separate wasm instructions from memory files
